refactor: Update API routes to replace syncProAccounts with syncSubscribers and adjust modal alt text
