+++ apps/api/src/routes/guild/index.ts
@@ -2,11 +2,11 @@ import { Hono } from "hono";
 import secretMiddleware from "src/middlewares/secretMiddleware";
 import syncFollowersStanding from "./syncFollowersStanding";
 import syncHQScoreAccounts from "./syncHQScoreAccounts";
-import syncProAccounts from "./syncProAccounts";
+import syncSubscribers from "./syncSubscribers";
 
 const app = new Hono();
 
-app.get("/syncProAccounts", secretMiddleware, syncProAccounts);
+app.get("/syncSubscribers", secretMiddleware, syncSubscribers);
 app.get("/syncHQScoreAccounts", secretMiddleware, syncHQScoreAccounts);
 app.get("/syncFollowersStanding", secretMiddleware, syncFollowersStanding);
