+++ apps/api/src/routes/guild/syncSubscribers.ts
